transfer - project list API updates

+++ app/Services/CountriesServices.php
@@ -51,7 +51,7 @@ class CountriesServices
             'system_type' => $request['system_type'] ?? '',
             'costing_status' => "Pending",
             'fee' => (float)$request['fee'] ?? 0,
-            'bond' => (int)$request['bond'] ?? 0,
+            'bond' => (float)$request['bond'] ?? 0,
             'created_by'    => $request['created_by'] ?? 0,
             'modified_by'   => $request['created_by'] ?? 0,
             'status' => 1,
@@ -83,7 +83,7 @@ class CountriesServices
                 'system_type' => $request['system_type'] ?? $country['system_type'],
                 'costing_status' => $country['costing_status'],
                 'fee' => (float)$request['fee'] ?? $country['fee'],
-                'bond' => (int)$request['bond'] ?? $country['bond'],
+                'bond' => (float)$request['bond'] ?? $country['bond'],
                 'modified_by'   => $request['modified_by'] ?? $country['modified_by'],
                 'status' => $country['status']
             ]),
